Replace uses of int with long.

+++ source/commands.c
@@ -80,10 +80,10 @@ void print_weakrefs() {
 // Print the global environment.
 void print_env(bool print_hash) {
     struct binding * b;
-    for (int i = 0; i < HASHSIZE; ++i) {
+    for (long i = 0; i < HASHSIZE; ++i) {
 	if (print_hash && global_env[i] != NULL) {
 	    printf("---\n");
-	    printf("%d\n", i);
+	    printf("%ld\n", i);
 	    printf("---\n");
 	}
 	for (b = global_env[i]; b != NULL; b = b->next) {
